+++ tests/registry/test_overrides.py
@@ -1,14 +1,6 @@
 import pytest
 
-from picodi import (
-    Provide,
-    SingletonScope,
-    dependency,
-    init_dependencies,
-    inject,
-    registry,
-    shutdown_dependencies,
-)
+from picodi import Provide, SingletonScope, inject, registry
 
 
 def get_abc_settings() -> dict:
@@ -162,7 +154,7 @@ def test_can_use_dep_with_not_default_scope_class_in_override(closeable):
     def my_service(settings: dict = Provide(get_abc_settings)):
         return settings
 
-    @dependency(scope_class=SingletonScope)
+    @registry.set_scope(SingletonScope)
     def real_settings():
         yield {"real": "settings"}
         closeable.close()
@@ -173,7 +165,7 @@ def test_can_use_dep_with_not_default_scope_class_in_override(closeable):
 
     assert result == {"real": "settings"}
     assert closeable.is_closed is False
-    shutdown_dependencies()
+    registry.shutdown()
     assert closeable.is_closed is True
 
 
@@ -197,11 +189,11 @@ async def test_can_use_async_dep_with_not_default_scope_in_override_in_sync_cont
         return settings
 
     @registry.override(get_abc_settings)
-    @dependency(scope_class=SingletonScope)
+    @registry.set_scope(SingletonScope)
     async def real_settings():
         return {"real": "settings"}
 
-    await init_dependencies([real_settings])
+    await registry.init([real_settings])
 
     result = my_service()